requirements.txt
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
idna==2.9
pexpect==4.8.0
prompt-toolkit==3.0.3
ptyprocess==0.6.0
PyYAML==5.3
requests==2.23.0
urllib3==1.25.8
wcwidth==0.1.8
